requirements.txt
anyio==4.10.0
argon2-cffi==25.1.0
argon2-cffi-bindings==25.1.0
arrow==1.3.0
asgiref==3.9.1
asttokens==3.0.0
async-lru==2.0.5
attrs==25.3.0
babel==2.17.0
bcc==0.18.0
beautifulsoup4==4.13.4
black==25.1.0
bleach==6.2.0
blinker==1.9.0
Brlapi==0.8.3
cachelib==0.13.0
certifi==2020.6.20
cffi==1.17.1
chardet==4.0.0
charset-normalizer==3.4.3
chrome-gnome-shell==0.0.0
click==8.2.1
colorama==0.4.4
comm==0.2.3
command-not-found==0.3
cryptography==3.4.8
cupshelpers==1.0
dbus-python==1.2.18
debugpy==1.8.16
decorator==5.2.1
defer==1.0.6
defusedxml==0.7.1
distlib==0.4.0
distro==1.7.0
Django==5.2.5
exceptiongroup==1.3.0
executing==2.2.0
fail2ban==0.11.2
fastjsonschema==2.21.2
filelock==3.19.1
Flask==3.1.2
Flask-Caching==2.3.1
fqdn==1.5.1
gpg==1.16.0
greenlet==1.1.2
gyp==0.1
h11==0.16.0
hidpidaemon==18.4.6
httpcore==1.0.9
httplib2==0.20.2
httpx==0.28.1
idna==3.3
importlib-metadata==4.6.4
iniconfig==2.1.0
ipykernel==6.30.1
ipython==8.37.0
isoduration==20.11.0
isort==6.0.1
itsdangerous==2.2.0
jedi==0.19.2
jeepney==0.7.1
Jinja2==3.1.6
json5==0.12.1
jsonpointer==3.0.0
jsonschema==4.25.1
jsonschema-specifications==2025.4.1
jupyter-events==0.12.0
jupyter-lsp==2.2.6
jupyter_client==8.6.3
jupyter_core==5.8.1
jupyter_server==2.17.0
jupyter_server_terminals==0.5.3
jupyterlab==4.4.6
jupyterlab_pygments==0.3.0
jupyterlab_server==2.27.3
kernelstub==3.1.4
keyring==23.5.0
language-selector==0.1
lark==1.2.2
launchpadlib==1.10.16
lazr.restfulclient==0.14.4
lazr.uri==1.0.6
louis==3.20.0
macaroonbakery==1.3.1
MarkupSafe==3.0.2
matplotlib-inline==0.1.7
mistune==3.1.3
more-itertools==8.10.0
msgpack==1.0.3
mypy_extensions==1.1.0
nbclient==0.10.2
nbconvert==7.16.6
nbformat==5.10.4
nest-asyncio==1.6.0
netaddr==0.8.0
netifaces==0.11.0
notebook_shim==0.2.4
oauthlib==3.2.0
overrides==7.7.0
packaging==25.0
pandocfilters==1.5.1
parso==0.8.4
pathspec==0.12.1
pexpect==4.9.0
platformdirs==4.3.8
pluggy==1.6.0
pop-transition==1.1.2
prometheus_client==0.22.1
prompt_toolkit==3.0.51
protobuf==3.12.4
psutil==7.0.0
ptyprocess==0.7.0
pure_eval==0.2.3
pycairo==1.20.1
pycparser==2.22
pycups==2.0.1
pydbus==0.6.0
Pygments==2.19.2
PyGObject==3.42.1
pyinotify==0.9.6
PyJWT==2.3.0
pymacaroons==0.13.0
PyNaCl==1.5.0
pynvim==0.4.2
pyparsing==2.4.7
PyQt5==5.15.6
PyQt5-sip==12.9.1
pyRFC3339==1.1
PySocks==1.7.1
pytest==8.4.1
pytest-mock==3.14.1
python-apt==2.4.0+ubuntu4
python-dateutil==2.9.0.post0
python-debian==0.1.43+ubuntu1.1
python-dotenv==1.1.1
python-gnupg==0.4.8
python-json-logger==3.3.0
python-xlib==0.29
pytz==2022.1
pyxdg==0.27
PyYAML==5.4.1
pyzmq==27.0.2
referencing==0.36.2
repolib==2.2.1
repoman==1.4.0
requests==2.32.5
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rfc3987-syntax==1.1.0
rpds-py==0.27.0
screen-resolution-extra==0.0.0
SecretStorage==3.3.1
Send2Trash==1.8.3
sessioninstaller==0.0.0
six==1.16.0
sniffio==1.3.1
soupsieve==2.7
sqlparse==0.5.3
stack-data==0.6.3
stdeb==0.10.0
systemd-python==234
terminado==0.18.1
tinycss2==1.4.0
tomli==2.2.1
tornado==6.5.2
traitlets==5.14.3
types-python-dateutil==2.9.0.20250822
typing_extensions==4.14.1
ubuntu-drivers-common==0.0.0
ufw==0.36.1
uri-template==1.3.0
urllib3==1.26.5
uv==0.8.13
virtualenv==20.34.0
wadllib==1.3.6
wcwidth==0.2.13
webcolors==24.11.1
webencodings==0.5.1
websocket-client==1.8.0
Werkzeug==3.1.3
xdg==5
xkit==0.0.0
zipp==1.0.0
